Save products from prom to json

+++ Promp/Promp/Models/Prom/PriceModel.cs
@@ -6,6 +6,6 @@ namespace Promp.Prom.Models
     {
         public double Price { get; set; }
         [JsonProperty(PropertyName = "Minimum_order_quantity")]
-        public int MinimumOrderQuantity { get; set; }
+        public float MinimumOrderQuantity { get; set; }
     }
 }
